Remove instructions to modify the nodecg-io build system from README
Adding skates-bundles to the nodecg-io build script is not needed any more as of https://github.com/codeoverflow-org/nodecg-io/pull/394. The nodecg-io build script now checks which packages are added to the npm workspace to figure out which packages it needs to build.
Note: current installs will work just fine after pulling. Just make sure to overwrite the local changes of the build script with the updated version when pulling.